Added username information in test to pass it
