+++ test/debate.js
@@ -2,6 +2,8 @@ import {SocketConfig} from '../src/conf/config.js'
 import {DebateManager} from "../src/debatemanager.js";
 import io from 'socket.io-client'
 import chai from 'chai';
+import {DataBaseManager} from "../src/database/DatabaseManager.js";
+import {Administrator} from "../src/database/modele/Users.js";
 
 const expect = chai.expect;
 const should = chai.should();
@@ -9,10 +11,18 @@ const should = chai.should();
 describe('Debate test', () => {
     let debateManager;
     let admin;
+    const db = new DataBaseManager();
 
     before(() => {
         debateManager = new DebateManager();
         debateManager.start();
+        db.start();
+        const admin = new Administrator({
+            login: 'admin',
+            password: 'pass'
+        });
+        let userSaved = admin.save();
+        console.log("administrator saved : ", userSaved);
     });
 
     describe("New debate", () => {
